Fix flowMatrix rejection by the read-only guard; make the fake enforce the client contract (CHAOS-4506)
P1: the registered flowMatrix route returned EMPTY for every dimension.

The pinned dev-health-go v0.4.0 client rejects a statement unless it
passes validateReadOnlyStatement (clickhouse/client.go:185-197): no ';',
first token must be SELECT, no external-table-function or INTO
OUTFILE/DUMPFILE. Four of six flowMatrix templates led with WITH -- TEAM
nodes and edges directly, REPO and WORK_TYPE edges via the shared
`WITH enriched AS (` prefix -- so they were rejected before reaching
ClickHouse. ExecuteFlowMatrix fails if either half fails, so ALL THREE
dimensions returned nothing, and resolveFlowMatrix swallowed the
rejection into an empty result with no GraphQL error: a totally broken
feature that looked like an org with no data.

Restructured all four to the inlined-subquery JOIN shape from Lane A's
workgraph/scope.go, which hit the identical guard live in its slot.

The real fix is the fake, not the four templates. A fake with no
statement guard cannot fail on this class, so the next WITH-leading
template would reintroduce it silently. clientcontract_test.go now
replicates validateReadOnlyStatement exactly and routingFakeClient runs
it on every query, so a rejected statement fails in `go test` instead of
degrading in production. Enumerated deliberately from client.go rather
than discovered one outage at a time.

Two more instances of the same kinder-than-reality fake, both found by
the same question:

- graphqldate.Date has no Scan method, so it is not an sql.Scanner and
  the native Date.ScanRow rejects it -- every non-empty timeseries result
  failed. The fake accepted *graphqldate.Date, a type the driver does
  not. Now scans time.Time and converts with graphqldate.New
  (reviewedges.go:152 precedent), and the fake mirrors the driver's
  accepted destination set.
- ClickHouse returns UInt64 for the SUM()-based measures and Float64 for
  AVG/ratio, and the driver errors rather than converting. breakdown and
  timeseries scanned float64 unconditionally; sankey scanned uint64
  unconditionally while sharing queryNodes/queryEdges with flowMatrix.
  Every value expression is now coerced with toFloat64 in SQL so one
  scan type serves all measures -- the same coercion Python does one
  layer later with float(row["value"]). The fake's checked assertions
  now report a type mismatch instead of panicking.

Also restores sankey's THREE-state useInvestment. analytics.py:634-636
passes batch.use_investment through UNWRAPPED, not bool(...), so when
both levels are unset the value stays None and _get_context_params
(compiler.py:152-155) auto-routes THEME/SUBCATEGORY/WORK_TYPE to the
investment path. resolveSankey took an already-collapsed bool, so unset
became false and the auto-route could never fire: Go silently applied
non-investment semantics, or emitted structurally invalid WORK_TYPE SQL
since investment_metrics_daily has no work_item_type column. The comment
above the defect cited analytics.py:554's bool(batch.use_investment) --
true for timeseries/breakdown, not for sankey. A timeseries fact
generalised to sankey is how the bug got written.

Verified by removal check, each restored and diffed: a WITH-leading
template, an embedded semicolon, the collapsed tri-state, and reverting
the Date scan each turn a test red.

Found by codex round 3, the first correctly-scoped round -- rounds 1 and
2 ran in the main ops checkout and reviewed a different tree.

+++ cmd/query-api/internal/analytics/sankey_test.go
@@ -179,9 +179,9 @@ func (f *concurrentFakeClient) Query(_ context.Context, statement string, _ []cl
 func TestExecuteSankeyQueries_MultiEdgeSuccess(t *testing.T) {
 	client := &concurrentFakeClient{
 		responses: map[string]*fakeRowScanner{
-			"'TEAM' AS dimension":        {rows: [][]any{{"TEAM", "team-a", uint64(3)}}},
-			"'TEAM' AS source_dimension": {rows: [][]any{{"TEAM", "REPO", "team-a", "repo-x", uint64(1)}}},
-			"'REPO' AS source_dimension": {rows: [][]any{{"REPO", "WORK_TYPE", "repo-x", "bug", uint64(2)}}},
+			"'TEAM' AS dimension":        {rows: [][]any{{"TEAM", "team-a", float64(3)}}},
+			"'TEAM' AS source_dimension": {rows: [][]any{{"TEAM", "REPO", "team-a", "repo-x", float64(1)}}},
+			"'REPO' AS source_dimension": {rows: [][]any{{"REPO", "WORK_TYPE", "repo-x", "bug", float64(2)}}},
 		},
 	}
 	nodesQ := []compiledQuery{{sql: "SELECT 'TEAM' AS dimension, ..."}}
